Mentor and Last meeting extracted to components
